feat(sql-query): Implementing sql query functionality
