+++ ui/src/api-logic/graphql/queries/sql-queries.ts
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client'
+
+const EXECUTE_SQL = gql`
+  query executeSQL($sql: String!) {
+    execSQL(query: $sql)
+  }
+`
+
+export { EXECUTE_SQL }
